requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
playwright==1.44.0
anthropic==0.28.0
cryptography==42.0.8
python-multipart==0.0.9
pydantic==2.7.1
